+++ udnssdk.go
@@ -72,6 +72,8 @@ type Client struct {
 	RRSets *RRSetsService
 	// Tasks API
 	Tasks *TasksService
+	// Zones API
+	Zones *ZonesService
 }
 
 // NewClient returns a new ultradns API client.
@@ -98,6 +100,7 @@ func NewClient(username, password, baseURL string) (*Client, error) {
 	c.Probes = &ProbesService{client: c}
 	c.RRSets = &RRSetsService{client: c}
 	c.Tasks = &TasksService{client: c}
+	c.Zones = &ZonesService{client: c}
 	return c, nil
 }
 
@@ -121,6 +124,7 @@ func newStubClient(username, password, baseURL, clientID, clientSecret string) (
 	c.Probes = &ProbesService{client: c}
 	c.RRSets = &RRSetsService{client: c}
 	c.Tasks = &TasksService{client: c}
+	c.Zones = &ZonesService{client: c}
 	return c, nil
 }