+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.3.1] - 2017-03-03
 ### Changed
+- ZoneKey: add, wrap users of string Zones
 - Client.NewRequest: shallow-copy BaseURL to avoid retaining modifications
 
 ## [1.3.0] - 2017-02-28